requirements.txt
# Web Backend
flask==3.0.3
flask-cors==4.0.1
flasgger==0.9.7.1
anthropic==0.34.2
groq==0.11.0
python-dotenv==1.0.1
Pillow>=10.0.0

# Database
sqlalchemy>=2.0.0

# Telegram Bot - TelebotAPI
pyTelegramBotAPI==4.22.1
requests>=2.32.0

# Telegram Bot - Aiogram
aiogram==3.13.1
aiohttp>=3.10.0
aiosqlite>=0.20.0

# Neural Network
tensorflow>=2.20.0
numpy>=2.1.0
matplotlib>=3.9.0
scikit-learn>=1.5.0
seaborn>=0.13.0
opencv-python-headless>=4.10.0